product-approach/infrastructure/requirements.txt
aws-cdk-lib>=2.90.0
constructs>=10.2.70
python-dotenv>=1.0.0